+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.8
 
+* Added world size to the system command
 * Added security manager for more efficient blocking checks
 * Added combined graphs example: /graph cpu heap threads